Refactor InstInterpreterCases
