+++ MeTTaIL/src/main/scala/io/f1r3fly/mettail/InstInterpreterCases.scala
@@ -3,64 +3,12 @@ package io.f1r3fly.mettail
 import metta_venus.Absyn._
 import metta_venus.PrettyPrinter
 import scala.jdk.CollectionConverters._
-import io.f1r3fly.mettail.ModuleProcessor
 
 object InstInterpreterCases {
-
-  trait LabelHelpers {
-    def labelsInAST(ast: AST): Set[String]
-    def labelsInEquation(eq: Equation): Set[String]
-    def labelsInRewrite(rw: Rewrite): Set[String]
-    def sequence[E, A](eithers: List[Either[E, A]]): Either[E, List[A]]
-  }
-
-  def labelToString(l: Label): String = l match {
-    case id: Id => id.ident_
-    case _: Wild => "_"
-    case _: ListE => "[]"
-    case _: ListCons => "(:)"
-    case _: ListOne => "(:[])"
-  }
-
-  private object BasePresOps {
-    import scala.jdk.CollectionConverters._
-
-    // Create an empty instance by calling the no-argument constructors.
-    def empty: BasePres = {
-      val listCat = new metta_venus.Absyn.ListCat()
-      val listDef = new metta_venus.Absyn.ListDef()
-      val listEquation = new metta_venus.Absyn.ListEquation()
-      val listRewriteDecl = new metta_venus.Absyn.ListRewriteDecl()
-      new BasePres(listCat, listDef, listEquation, listRewriteDecl)
-    }
-
-    // copyPres creates new list instances and adds the updated elements.
-    def copyPres(
-      pres: BasePres,
-      listcat: Option[List[Cat]] = None,
-      listdef: Option[List[Def]] = None,
-      listequation: Option[List[Equation]] = None,
-      listrewritedecl: Option[List[RewriteDecl]] = None
-    ): BasePres = {
-      val newListCat = new metta_venus.Absyn.ListCat()
-      val newListDef = new metta_venus.Absyn.ListDef()
-      val newListEquation = new metta_venus.Absyn.ListEquation()
-      val newListRewriteDecl = new metta_venus.Absyn.ListRewriteDecl()
-
-      val cats = listcat.getOrElse(pres.listcat_.asScala.toList)
-      newListCat.addAll(cats.asJava)
-      val defs = listdef.getOrElse(pres.listdef_.asScala.toList)
-      newListDef.addAll(defs.asJava)
-      val equations = listequation.getOrElse(pres.listequation_.asScala.toList)
-      newListEquation.addAll(equations.asJava)
-      val rewrites = listrewritedecl.getOrElse(pres.listrewritedecl_.asScala.toList)
-      newListRewriteDecl.addAll(rewrites.asJava)
-
-      new BasePres(newListCat, newListDef, newListEquation, newListRewriteDecl)
-    }
-  }
-
   import BasePresOps._
+  import InstInterpreterHelpers._
+  import InstInterpreterHelpers.addEquationsHelpers._
+  import ModuleProcessor._
 
   def handleEmpty(): Either[String, BasePres] =
     Right(empty)
@@ -165,24 +113,6 @@ object InstInterpreterCases {
                      listrewritedecl = Some(diffRewrites.toList))
   }
 
-  private def replaceCats(oldCat: Cat, newCat: Cat, listItem: ListItem): ListItem = {
-    val javaList = listItem.asScala.map {
-      case t: Terminal => t
-      case nt: NTerminal => if (nt.cat_ == oldCat) new NTerminal(newCat) else nt
-      case ant: AbsNTerminal => if (ant.cat_ == oldCat) new AbsNTerminal(ant.ident_, newCat) else ant
-      case bnt: BindNTerminal => if (bnt.cat_ == oldCat) new BindNTerminal(bnt.ident_, newCat) else bnt
-    }.toSeq.asJava
-    val newListItem = new ListItem()
-    newListItem.addAll(javaList)
-    newListItem
-  }
-
-  // Helper: Update a definition (Def). For Rule definitions, apply updateRule.
-  private def updateDef(d: Def, oldCat: Cat, newCat: Cat): Def = d match {
-    case rule: Rule => new Rule(rule.label_, newCat, replaceCats(oldCat, newCat, rule.listitem_))
-    case other   => other
-  }
-
   def handleAddExports(interpreter: InstInterpreter, env: List[(String, BasePres)], inst: TheoryInstAddExports): Either[String, BasePres] =
     interpreter.interpret(env, inst.theoryinst_).flatMap { basePres =>
       // Process each export instruction sequentially.
@@ -210,10 +140,6 @@ object InstInterpreterCases {
       }
     }
 
-  def nonTerminals(items: ListItem): Seq[Item] = {
-    items.asScala.toSeq.filter(item => !item.isInstanceOf[Terminal])
-  }
-
   def handleAddReplacements(interpreter: InstInterpreter,
                             env: List[(String, BasePres)],
                             inst: TheoryInstAddReplacements): Either[String, BasePres] = {
@@ -224,13 +150,10 @@ object InstInterpreterCases {
         inst.listreplacement_.asScala.toList.collect { case s: SimpleRepl => s }
 
       // Helper function to convert an IntList (defined in Absyn/IntList.java) into a Scala List[Int].
-      def convertIntList(il: IntList): List[Int] =
-        il.accept(new IntList.Visitor[List[Int], Unit] {
-          override def visit(p: Ints, arg: Unit): List[Int] = {
-            // p.listinteger_ is a ListInteger (a LinkedList<Integer>)
-            p.listinteger_.asScala.toList.map(identity)
-          }
-        }, ())
+      def convertIntList(intList: IntList): List[Int] = intList match {
+        case ints: Ints =>
+          ints.listinteger_.asScala.toList.map(_.intValue())
+      }
 
       // Process each replacement sequentially.
       replacements.foldLeft[Either[String, BasePres]](Right(basePres)) { (accEither, s) =>
@@ -277,7 +200,7 @@ object InstInterpreterCases {
                             case r: Rule if r.label_.toString == s.label_.toString => s.def_
                             case other => other
                           }
-                          val updatedPres = InstInterpreterCases.BasePresOps.copyPres(currentPres, listdef = Some(newDefs))
+                          val updatedPres = copyPres(currentPres, listdef = Some(newDefs))
                           Right(updatedPres)
                         }
                       }
@@ -321,186 +244,6 @@ object InstInterpreterCases {
       }
     }
 
-  private object addEquationsHelpers {
-    sealed trait CatOfASTResult
-
-    case class LabelNotFound(label: Label) extends CatOfASTResult
-    case class Var(varName: String) extends CatOfASTResult
-    case class Concrete(cat: Cat) extends CatOfASTResult
-
-    def printCatOfASTResult(c: CatOfASTResult) = {
-      c match {
-        case LabelNotFound(l) => s"LabelNotFound(${PrettyPrinter.print(l)})"
-        case Var(v) => s"Var($v)"
-        case Concrete(c) => s"Concrete(${PrettyPrinter.print(c)})"
-      }
-    }
-
-    def catOfAST(ast: AST, defs: Map[Label, Rule]): CatOfASTResult = {
-      ast match {
-        case astVar: ASTVar => Var(astVar.ident_)
-        case astSExp: ASTSExp => defs.get(astSExp.label_) match {
-          case None => LabelNotFound(astSExp.label_)
-          case Some(rule) => Concrete(rule.cat_)
-        }
-        case astSubst: ASTSubst => catOfAST(astSubst.ast_1, defs) match {
-          case lnf: LabelNotFound => lnf
-          case Var(v) if astSubst.ident_ == v => catOfAST(astSubst.ast_2, defs)
-          case other => other
-        }
-      }
-    }
-
-    def freeVarsInAST(ast: AST): Set[String] = {
-      ast match {
-        case astVar: ASTVar => Set(astVar.ident_)
-        case astSExp: ASTSExp => astSExp.listast_.asScala.flatMap(freeVarsInAST).toSet
-        case astSubst: ASTSubst => 
-          freeVarsInAST(astSubst.ast_1).filterNot(_ == astSubst.ident_) ++ freeVarsInAST(astSubst.ast_2)
-      }
-    }
-
-    def listDefToMap(listDef: ListDef): Map[Label, Rule] = {
-      listDef.asScala.collect {
-        case rule: Rule => rule.label_ -> rule
-      }.toMap
-    }
-
-    sealed trait CatOfIdentInASTResult
-
-    case class COIIAError(err: String) extends CatOfIdentInASTResult
-    case object COIIANotInAST extends CatOfIdentInASTResult
-    case class COIIAVar(varName: String) extends CatOfIdentInASTResult
-    case class COIIAConcrete(cat: Cat) extends CatOfIdentInASTResult
-
-    def catOfIdentInAST(
-      ident: String,
-      defs: Map[Label, Rule],
-      context: Option[Cat],
-      ast: AST
-    ): CatOfIdentInASTResult = {
-      ast match {
-        case astSubst: ASTSubst => handleASTSubst(ident, defs, context, astSubst)
-        case astVar: ASTVar     => handleASTVar(ident, defs, context, astVar)
-        case astSExp: ASTSExp   => handleASTSExp(ident, defs, context, astSExp)
-        case _                  => COIIAError(s"Unknown AST type: ${PrettyPrinter.print(ast)}")
-      }
-    }
-
-    def optCatFromItem(item: Item): Option[Cat] = {
-      item match {
-        case nt: NTerminal => Some(nt.cat_)
-        case ant: AbsNTerminal => Some(ant.cat_)
-        case bnt: BindNTerminal => Some(new IdCat(bnt.ident_))
-        case _: Terminal => None
-      }
-    }
-
-    def handleASTSExp(ident: String, defs: Map[Label, Rule], context: Option[Cat], astSExp: ASTSExp): CatOfIdentInASTResult = {
-      // match length with rule
-      val optRule = defs.get(astSExp.label_)
-      optRule match {
-        case None => COIIAError(s"No rule with label ${PrettyPrinter.print(astSExp.label_)}")
-        case Some(rule) => {
-          val filtered = nonTerminals(rule.listitem_)
-          val children = astSExp.listast_.asScala
-          if (filtered.length != children.length) {
-            COIIAError(s"Length mismatch with label ${PrettyPrinter.print(astSExp.label_)}")
-          } else {
-            // match category of rule with context
-            context match {
-              case Some(ctxCat) if rule.cat_ != ctxCat => 
-                COIIAError(s"Label ${PrettyPrinter.print(astSExp.label_)}'s"
-                           + s" category ${PrettyPrinter.print(rule.cat_)} doesn't"
-                           + s" match context ${PrettyPrinter.print(ctxCat)}")
-              case _ => {
-                // fold over children
-                val childCats = children.zipWithIndex.map {
-                  case (child, idx) => {
-                    val ctx = optCatFromItem(filtered(idx))
-                    catOfIdentInAST(ident, defs, ctx, child)
-                  }
-                }.toSeq
-                val result = childCats.foldLeft[CatOfIdentInASTResult](
-                  COIIANotInAST
-                ) {
-                  // An error in any subtree propagates up.
-                  case (COIIAError(err), _) => COIIAError(err)
-                  case (_, COIIAError(err)) => COIIAError(err)
-                  // If it's not found in one subtree, defer to any other result.
-                  case (COIIANotInAST, other) => other
-                  case (other, COIIANotInAST) => other
-                  case (COIIAConcrete(x), COIIAConcrete(y)) => if (x == y) {
-                    COIIAConcrete(x)
-                  } else {
-                    COIIAError(s"Identifier $ident has mismatched categories"
-                               + s" ${PrettyPrinter.print(x)} and ${PrettyPrinter.print(y)}"
-                               + s" in ${PrettyPrinter.print(astSExp)}")
-                  }
-                  case pair => COIIAError(s"Unexpected pairing in fold: $pair")
-                }
-                result
-              }
-            }
-          }
-        }
-      }
-    }
-
-    def handleASTVar(ident: String, defs: Map[Label, Rule], context: Option[Cat], astVar: ASTVar): CatOfIdentInASTResult = {
-      if (ident == astVar.ident_) {
-        context match {
-          case None => COIIAVar(ident)
-          case Some(ctx) => COIIAConcrete(ctx)
-        }
-      } else {
-        COIIANotInAST
-      }
-    }
-
-    def handleASTSubst(ident: String, defs: Map[Label, Rule], context: Option[Cat], astSubst: ASTSubst): CatOfIdentInASTResult = {
-      catOfIdentInAST(ident, defs, context, findAndReplace(astSubst.ast_2, astSubst.ident_, defs)(astSubst.ast_1))
-    }
-
-    // Precondition: any label in the AST has to appear in defs.
-    def findAndReplace(replacement: AST, ident: String, defs: Map[Label, Rule])(ast: AST): AST = {
-      ast match {
-        case astVar: ASTVar => if (astVar.ident_ == ident) replacement else astVar
-        case astSExp: ASTSExp => {
-          val optRule = defs.get(astSExp.label_)
-          optRule match {
-            case Some(rule) => {
-              val filtered = nonTerminals(rule.listitem_)
-              val javaList = filtered.zip(astSExp.listast_.asScala).map {
-                case (item, ast) => item match {
-                  // In the usual case, we recursively replace
-                  case nt: NTerminal => findAndReplace(replacement, ident, defs)(ast)
-
-                  // If the bound variable shadows the one being replaced, don't recurse; otherwise, do
-                  case ant: AbsNTerminal =>
-                    if (ant.ident_ == ident) ast else findAndReplace(replacement, ident, defs)(ast)
-
-                  // Never replace in a BindNTerminal
-                  case bnt: BindNTerminal => ast
-                }
-              }.toSeq.asJava
-              val newListAST = new ListAST()
-              newListAST.addAll(javaList)
-              new ASTSExp(astSExp.label_, newListAST)
-            }
-          }
-        }
-        case astSubst: ASTSubst => {
-          findAndReplace(replacement, ident, defs)(
-            findAndReplace(astSubst.ast_2, astSubst.ident_, defs)(astSubst.ast_1)
-          )
-        }
-      }
-    }
-  }
-
-  import addEquationsHelpers._
-
   def handleAddEquations(interpreter: InstInterpreter,
                           env: List[(String, BasePres)],
                           inst: TheoryInstAddEquations): Either[String, BasePres] = {
@@ -524,18 +267,18 @@ object InstInterpreterCases {
                 val leftCat = catOfAST(eqn.ast_1, defs)
                 val rightCat = catOfAST(eqn.ast_2, defs)
                 val sameCategory: Either[String, Cat] = leftCat match {
-                  case LabelNotFound(label) => Left(s"Label $label not found in equation ${PrettyPrinter.print(eqn)}")
-                  case Var(leftVarName) => rightCat match {
-                    case LabelNotFound(label) => Left(s"Label $label not found in equation ${PrettyPrinter.print(eqn)}")
-                    case Var(rightVarName) =>
+                  case COALabelNotFound(label) => Left(s"Label $label not found in equation ${PrettyPrinter.print(eqn)}")
+                  case COAVar(leftVarName) => rightCat match {
+                    case COALabelNotFound(label) => Left(s"Label $label not found in equation ${PrettyPrinter.print(eqn)}")
+                    case COAVar(rightVarName) =>
                       Left(s"Cannot determine categories of variables $leftVarName and $rightVarName"
                            + s" in equation ${PrettyPrinter.print(eqn)}")
-                    case Concrete(rightConcreteCat) => Right(rightConcreteCat)
+                    case COAConcrete(rightConcreteCat) => Right(rightConcreteCat)
                   }
-                  case Concrete(leftConcreteCat) => rightCat match {
-                    case LabelNotFound(label) => Left(s"Label $label not found in equation ${PrettyPrinter.print(eqn)}")
-                    case Var(rightVarName) => Right(leftConcreteCat)
-                    case Concrete(rightConcreteCat) => if (leftConcreteCat == rightConcreteCat) {
+                  case COAConcrete(leftConcreteCat) => rightCat match {
+                    case COALabelNotFound(label) => Left(s"Label $label not found in equation ${PrettyPrinter.print(eqn)}")
+                    case COAVar(rightVarName) => Right(leftConcreteCat)
+                    case COAConcrete(rightConcreteCat) => if (leftConcreteCat == rightConcreteCat) {
                       Right(leftConcreteCat)
                     } else {
                       Left(s"Categories of the sides differ (${PrettyPrinter.print(leftConcreteCat)}"
@@ -626,7 +369,6 @@ object InstInterpreterCases {
       // Extract the new rewrite declarations from the instruction.
       val newRewrites = inst.listrewritedecl_.asScala.toList
 
-      // First, perform the label check (as in handleConj).
       val allowedLabels: Set[String] = basePres.listdef_.asScala.collect {
         case rule: Rule => labelToString(rule.label_)
       }.toSet
@@ -696,7 +438,7 @@ object InstInterpreterCases {
   def handleCtor(interpreter: InstInterpreter, env: List[(String, BasePres)],
                  resolvedModules: Map[String, Module], currentModulePath: String,
                  ctor: TheoryInstCtor): Either[String, BasePres] = {
-    ModuleProcessor.resolveDottedPath(resolvedModules, currentModulePath, ctor.dottedpath_) match {
+    resolveDottedPath(resolvedModules, currentModulePath, ctor.dottedpath_) match {
       case Left(error) => Left(error)
       case Right((modulePath, theoryDecl)) => theoryDecl match {
         case baseDecl: BaseTheoryDecl =>
